correct warning from Windows build about size_t to int conversion

+++ arangosh/Import/SenderThread.h
@@ -56,7 +56,7 @@ class SenderThread final : public arangodb::Thread {
   //////////////////////////////////////////////////////////////////////////////
 
   void sendData(std::string const& url, basics::StringBuffer* sender,
-                int lowLine = 0, int highLine = 0);
+                size_t lowLine = 0, size_t highLine = 0);
 
   bool hasError();
   /// Ready to start sending
@@ -81,8 +81,8 @@ class SenderThread final : public arangodb::Thread {
   bool _hasError;
   bool _idle;
   bool _ready;
-  int _lowLineNumber;
-  int _highLineNumber;
+  size_t _lowLineNumber;
+  size_t _highLineNumber;
 
   ImportStatistics* _stats;
   std::string _errorMessage;
